refactor(shared-data): add ot3_standard deck
Adds a definition for the OT-3 standard deck (and the trash labware).
This will likely not be the final definition - it will need to grow more
complex since the trash, for instance, is likely movable, and also
prototypes will change - but it is accurate for and tested on current
hardware.

Co-Developed-By: Laura Cox <laura@opentrons.com>