refactor(api): add function to determine deck
The type of deck used on the current robot is currently determined
piecemeal around the system by every piece of code that needs to know
the deck. With the deck beginning to change more frequently and
fundamentally, it will be useful to centralize these places. A new
function in the same place that stores the standard deck can accomplish
this.

Co-Developed-By: Laura Cox <laura@opentrons.com>