refactor(api): ot3 gets its own HC API
With a mypy protocol defining the hardware API and some behavioral
changes that require alterations above the level of the controller, it's
both a better and safe idea to make the OT3 use a different hardware
control object rather than just a different controller. This PR removes
the OT3-specific changes and use of the OT3 controller from
hardware_control.api and puts them in hardware_control.ot3api.